[JIT] Fix bug in get_annotation_str for ast.Subscript
**Summary**
This commit fixes a bug in the handling of `ast.Subscript` inside
`get_annotation_str`. `annotation.value` (which contains the AST node
representing the container name) should also be processed using
`get_annotation_str`.

**Test Plan**
This commit adds a unit test to `TestClassType` based on the test case
from the issue that reported this bug.

**Fixes**
This commit fixes #47570.

ghstack-source-id: 7bea13231c41bfb4f6ea242ca9b861240ef7af2f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48741