[Gradient Compression] Add an index field to GradBucket for PowerSGD
Add an index field to GradBucekt, so error_dict is keyed by this index instead of the hashcode of input tensor.

Howevever, sometimes the buckets can be rebuilt in the forward pass. In this case, the shape of the bucket with the same index will not be consistent with the one in the previous iteration, and hence the error tensor will be re--initialized as a zero tensor of the new shape.

Original PR issue: Investigate Applying PowerSGD to Communication Hook for Gradient Compression #47202

Differential Revision: [D25288496](https://our.internmc.facebook.com/intern/diff/D25288496/)

ghstack-source-id: 117719173
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48757